telethon_service/requirements.txt
aiogram==3.0.0b5
sqlalchemy==2.0.20
psycopg2-binary==2.9.9
python-dotenv==1.0.1